% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_cluster.R
\name{co_cluster}
\alias{co_cluster}
\title{Hierarchical clustering was performed to look for co-expression relationships between miRNAs}
\usage{
co_cluster(data, h, k)
}
\arguments{
\item{data}{A dataframe or matrix of expression profile with case-sample and control-sample.The list of gene names is the rowname of the dataframe or matrix.}

\item{h}{A numeric variable with heights where the tree should be cut.}

\item{k}{A numeric variable with the number of clusters that data is divided.K can be left out.}
}
\value{
Split results of features and cluster dendrogram.
}
\description{
Hierarchical clustering was performed to look for co-expression relationships between miRNAs
}
\examples{
co_cluster(data,h=0.4) or cluster(data,h=0.4,k=3)
}
