% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FChange.R
\name{FChange}
\alias{FChange}
\title{fold change}
\usage{
FChange(datacase, datacontrol)
}
\arguments{
\item{datacase}{A data frame in which the columns are samples, the rows are features.The colunms'names are the names of samples, and the rows'names are the names of features.}

\item{datacontrol}{A data frame in which the columns are samples, the rows are features.The colunms'names are the names of samples, and the rows'names are the names of features.}
}
\value{
a data frame with two columns which are genenames and fold change.
}
\description{
fold change
}
\examples{
 FChange(datacase,datacontrol)
}
