% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAM_Analysis.R
\name{SAM_Analysis}
\alias{SAM_Analysis}
\title{Compute the value of sam between case and control,and find the significant miRNAs in train set}
\usage{
SAM_Analysis(datacase, datacontrol, p)
}
\arguments{
\item{datacase}{A dataframe of expression profile which the samples are case-sample.The list of gene names is the first column of the dataframe.}

\item{datacontrol}{A dataframe of expression profile which the samples are control-sample.The list of gene names is the first column of the dataframe.}

\item{p}{A numeric variable that controls the threshold value of differentially expression of genes.}

\item{delta}{A numeric variable that adjusts the false discovery rate(FDR)}
}
\value{
A dataframe of feature selection set which is consisted of genes with p value less than or equal to p and the expression value of them.
}
\description{
Compute the value of sam between case and control,and find the significant miRNAs in train set
}
\examples{
SAM_Analysis(datacase,datacontrol,p=0.05)
}
