% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{Sampling}
\alias{Sampling}
\title{Using two dataframes to create sample sets.It will return 3*5*count training sets, 3*5*count test sets and 3*5*count evaluation sets in your working space.}
\usage{
Sampling(datacase, datacontrol, count)
}
\arguments{
\item{datacase}{A dataframe of expression profile which the samples are case-sample.The list of gene names is the first column of the dataframe.}

\item{datacontrol}{A dataframe of expression profile which the samples are control-sample.The list of gene names is the first column of the dataframe.}

\item{count}{A numeric variable that controls the number of random samples.}
}
\value{
3*5*count training sets, 3*5*count test sets and 3*5*count evaluation sets in your working space.
}
\description{
Using two dataframes to create sample sets.It will return 3*5*count training sets, 3*5*count test sets and 3*5*count evaluation sets in your working space.
}
\examples{
Sampling(data1,data2,100)#data1 is a data.frame of expression profile which the samples are case-sample,data2 is also a data.frame of expression profile which the samples are control-sample
}
